FIX LAUNCH ERROR on V17 Krypton
Couldn't get this to run on a Mac on Kodi 17 as it had the incorrect path to the User Profile directory and the Addon_data folder. If you update the references that are incorrectly translating the path to the user profile's addon_data folder as special://home/addons when the correct path to translate should be: special://profile/addon_data so I corrected and this works on my Mac running Kodi 17 and uses multiple profiles. I can imagine this wouldn't maybe cause the addon to fail under dif circumstances but it wouldn't even load with this incorrect path.

"special://profile/addon_data/{0}/".format(addon_id)